Cleanup app/mod
