+++ src/app/opt/mod.rs
@@ -2,3 +2,4 @@ pub mod auth;
 pub mod endpoint;
 pub mod patch;
 pub mod yuse;
+pub mod to_value;
